Rebase on master and refactor
